requirements.txt
# Base mathematical
cython
numpy
scipy
pandas

# Machine learning
scikit-learn
mlxtend
vowpalwabbit
## Figure out from which distribution your data comes from
fitter
## Approximate Nearest Neighbors Oh Yeah
annoy

# ipython/juptyer
jupyter
ipython
ipywidgets

# Plotting
matplotlib
plotly
graphviz
pydot

# Image
pillow
scikit-image
opencv-python

# Text mining
nltk
gensim
sklearn

# Keras-related; Avoid Keras 2.0.9
theano
Keras==2.0.8
keras-vis

# File processing
lxml
pyyaml
ujson

# Reproducible experiments
sacred

# Local Interpretable Model-Agnostic Explanations
lime

# Profiling
pprofile
line_profiler

# HDF5
tables
h5py

# Parallel processing
sharedmem

# Progress meter
tqdm
keras-tqdm

# Virtual Environments
virtualenv

# Jupyter Extensions
jupyter_contrib_nbextensions
jupyterlab
